Renamed interface.h to widget.h (issue #10).
