+++ Libraries/Widgets/logo.cpp
